refactor: Test를 위한 클래스 Test pakage로 이동
