--- src/main/java/racing/domain/car/TestCarForwardBehavior.java
@@ -1,10 +1,0 @@
-package racing.domain.car;
-
-public class TestCarForwardBehavior implements CarMoveBehavior{
-    private final boolean GO = true;
-
-    @Override
-    public boolean action(){
-        return true;
-    }
-}
