--- src/main/java/racing/domain/car/TestCarStopBehavior.java
@@ -1,10 +1,0 @@
-package racing.domain.car;
-
-public class TestCarStopBehavior implements CarMoveBehavior{
-    private final boolean STOP = false;
-
-    @Override
-    public boolean action(){
-        return STOP;
-    }
-}
