+++ src/main/java/racing/domain/test/TestCarStopBehavior.java
@@ -1,0 +1,12 @@
+package racing.domain.test;
+
+import racing.domain.car.CarMoveBehavior;
+
+public class TestCarStopBehavior implements CarMoveBehavior {
+    private final boolean STOP = false;
+
+    @Override
+    public boolean action(){
+        return STOP;
+    }
+}
